strategy-engine/requirements.txt
akshare>=1.14.0
FinMind>=1.6.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tabulate>=0.9.0
